Broken out CacheMap so that its promises and values
