+++ src/main/java/org/dataloader/BatchLoaderContextProvider.java
@@ -1,5 +1,7 @@
 package org.dataloader;
 
+import org.dataloader.annotations.PublicSpi;
+
 /**
  * A BatchLoaderContextProvider is used by the {@link org.dataloader.DataLoader} code to
  * provide overall calling context to the {@link org.dataloader.BatchLoader} call.  A common use
