+++ src/main/java/org/dataloader/BatchLoaderEnvironment.java
@@ -1,5 +1,7 @@
 package org.dataloader;
 
+import org.dataloader.annotations.PublicApi;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
@@ -7,6 +9,7 @@ import java.util.List;
 import java.util.Map;
 
 import static java.util.Objects.nonNull;
+import static java.util.Objects.requireNonNull;
 
 /**
  * This object is passed to a batch loader as calling context.  It could contain security credentials
@@ -78,8 +81,8 @@ public class BatchLoaderEnvironment {
         }
 
         public <K> Builder keyContexts(List<K> keys, List<Object> keyContexts) {
-            nonNull(keys);
-            nonNull(keyContexts);
+            requireNonNull(keys);
+            requireNonNull(keyContexts);
 
             Map<Object, Object> map = new HashMap<>();
             List<Object> list = new ArrayList<>();
