+++ src/main/java/org/dataloader/DataLoader.java
@@ -1,5 +1,6 @@
 package org.dataloader;
 
+import org.dataloader.annotations.PublicApi;
 import org.dataloader.stats.Statistics;
 
 import java.util.List;
@@ -35,27 +36,137 @@ import java.util.concurrent.CompletableFuture;
  */
 @PublicApi
 public interface DataLoader<K, V> {
+    /**
+     * Requests to load the data with the specified key asynchronously, and returns a future of the resulting value.
+     * <p>
+     * If batching is enabled (the default), you'll have to call {@link org.dataloader.DataLoader#dispatch()} at a later stage to
+     * start batch execution. If you forget this call the future will never be completed (unless already completed,
+     * and returned from cache).
+     *
+     * @param key the key to load
+     * @return the future of the value
+     */
     CompletableFuture<V> load(K key);
 
+    /**
+     * Requests to load the data with the specified key asynchronously, and returns a future of the resulting value.
+     * <p>
+     * If batching is enabled (the default), you'll have to call {@link org.dataloader.DataLoader#dispatch()} at a later stage to
+     * start batch execution. If you forget this call the future will never be completed (unless already completed,
+     * and returned from cache).
+     * <p>
+     * The key context object may be useful in the batch loader interfaces such as {@link org.dataloader.BatchLoaderWithContext} or
+     * {@link org.dataloader.MappedBatchLoaderWithContext} to help retrieve data.
+     *
+     * @param key        the key to load
+     * @param keyContext a context object that is specific to this key
+     * @return the future of the value
+     */
     CompletableFuture<V> load(K key, Object keyContext);
 
+    /**
+     * Requests to load the list of data provided by the specified keys asynchronously, and returns a composite future
+     * of the resulting values.
+     * <p>
+     * If batching is enabled (the default), you'll have to call {@link org.dataloader.DataLoader#dispatch()} at a later stage to
+     * start batch execution. If you forget this call the future will never be completed (unless already completed,
+     * and returned from cache).
+     *
+     * @param keys the list of keys to load
+     * @return the composite future of the list of values
+     */
     CompletableFuture<List<V>> loadMany(List<K> keys);
 
+    /**
+     * Requests to load the list of data provided by the specified keys asynchronously, and returns a composite future
+     * of the resulting values.
+     * <p>
+     * If batching is enabled (the default), you'll have to call {@link org.dataloader.DataLoader#dispatch()} at a later stage to
+     * start batch execution. If you forget this call the future will never be completed (unless already completed,
+     * and returned from cache).
+     * <p>
+     * The key context object may be useful in the batch loader interfaces such as {@link org.dataloader.BatchLoaderWithContext} or
+     * {@link org.dataloader.MappedBatchLoaderWithContext} to help retrieve data.
+     *
+     * @param keys        the list of keys to load
+     * @param keyContexts the list of key calling context objects
+     * @return the composite future of the list of values
+     */
     CompletableFuture<List<V>> loadMany(List<K> keys, List<Object> keyContexts);
 
+    /**
+     * This will return an optional promise to a value previously loaded via a {@link #load(Object)} call or empty if not call has been made for that key.
+     * <p>
+     * If you do get a present CompletableFuture it does not mean it has been dispatched and completed yet.  It just means
+     * its at least pending and in cache.
+     * <p>
+     * If caching is disabled there will never be a present Optional returned.
+     * <p>
+     * NOTE : This will NOT cause a data load to happen.  You must called {@link #load(Object)} for that to happen.
+     *
+     * @param key the key to check
+     * @return an Optional to the future of the value
+     */
     Optional<CompletableFuture<V>> getIfPresent(K key);
 
+    /**
+     * This will return an optional promise to a value previously loaded via a {@link #load(Object)} call that has in fact been completed or empty
+     * if no call has been made for that key or the promise has not completed yet.
+     * <p>
+     * If you do get a present CompletableFuture it means it has been dispatched and completed.  Completed is defined as
+     * {@link java.util.concurrent.CompletableFuture#isDone()} returning true.
+     * <p>
+     * If caching is disabled there will never be a present Optional returned.
+     * <p>
+     * NOTE : This will NOT cause a data load to happen.  You must called {@link #load(Object)} for that to happen.
+     *
+     * @param key the key to check
+     * @return an Optional to the future of the value
+     */
     Optional<CompletableFuture<V>> getIfCompleted(K key);
 
 
+    /**
+     * Dispatches the queued load requests to the batch execution function and returns a promise of the result.
+     * <p>
+     * If batching is disabled, or there are no queued requests, then a succeeded promise is returned.
+     *
+     * @return the promise of the queued load requests
+     */
     CompletableFuture<List<V>> dispatch();
 
+    /**
+     * Dispatches the queued load requests to the batch execution function and returns both the promise of the result
+     * and the number of entries that were dispatched.
+     * <p>
+     * If batching is disabled, or there are no queued requests, then a succeeded promise with no entries dispatched is
+     * returned.
+     *
+     * @return the promise of the queued load requests and the number of keys dispatched.
+     */
     DispatchResult<V> dispatchWithCounts();
 
+    /**
+     * Normally {@link #dispatch()} is an asynchronous operation but this version will 'join' on the
+     * results if dispatch and wait for them to complete.  If the {@link CompletableFuture} callbacks make more
+     * calls to this data loader then the {@link #dispatchDepth()} will be &gt; 0 and this method will loop
+     * around and wait for any other extra batch loads to occur.
+     *
+     * @return the list of all results when the {@link #dispatchDepth()} reached 0
+     */
     List<V> dispatchAndJoin();
 
+    /**
+     * @return the depth of the batched key loads that need to be dispatched
+     */
     int dispatchDepth();
 
+    /**
+     * Gets the statistics associated with this data loader.  These will have been gather via
+     * the {@link org.dataloader.stats.StatisticsCollector} passed in via {@link DataLoaderOptions#getStatisticsCollector()}
+     *
+     * @return statistics for this data loader
+     */
     Statistics getStatistics();
