+++ src/main/java/org/dataloader/DataLoaderRegistry.java
@@ -8,6 +8,7 @@ import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.function.Function;
 
+import org.dataloader.annotations.PublicApi;
 import org.dataloader.stats.Statistics;
 
 /**
@@ -131,7 +132,7 @@ public class DataLoaderRegistry {
      */
     public int dispatchDepth() {
         int totalDispatchDepth = 0;
-        for (DataLoader dataLoader : getDataLoaders()) {
+        for (DataLoader<?,?> dataLoader : getDataLoaders()) {
             totalDispatchDepth += dataLoader.dispatchDepth();
         }
         return totalDispatchDepth;
